Rename myFiles to selectedFile
