+++ xroad-drive-ui/src/components/fileView/SharedFileList.vue
@@ -85,7 +85,7 @@
       // ...mapState('fileList', ['fileList'])
     },
     methods: {
-      ...mapActions('myFiles', ['updateSelectedFile']),
+      ...mapActions('selectedFile', ['updateSelectedFile']),
       // ...mapActions('fileList', ['fetchFileList']),
       clickRow(item, row) {
         if (this.selectedFile && row.isSelected) {
