+++ xroad-drive-ui/src/store/index.js
@@ -1,6 +1,6 @@
 import Vue from 'vue';
 import Vuex from 'vuex';
-import myFiles from './modules/myFiles';
+import selectedFile from './modules/selectedFile';
 import uploadFiles from './modules/uploadFiles';
 import fileList from './modules/fileList';
 
@@ -8,7 +8,7 @@ Vue.use(Vuex);
 
 export default new Vuex.Store({
   modules: {
-    myFiles,
+    selectedFile,
     uploadFiles,
     fileList
   }
