+++ xroad-drive-ui/src/store/modules/selectedFile.js
